Rationalize ConfigMap usage when paths (mounted files) are involved
This is done by removing the file walking and loading logic that existed
before and replacing it with an implementation that creates on
ConfigMapResource per specified path.
Reading and processing the content of the path is done in the exact
same way as the enableApi method does thus providing a consistent
experience.

The implementation is inspired by the following comment:
https://github.com/spring-cloud/spring-cloud-kubernetes/issues/230#issuecomment-420341743

Fixes: #230