Change the log level to prevent false-positives (#911)
The `WARN` log level is being abused here which can flood the production logs (usually you are not interested in any lower log levels there). It is totally fine to not define a separate configMap for each profile that is active in the application therefore it shouldn't be a warning. The same applies to any other form of externalized configuration.